python: accept .json as DATASTORE_PRIVATE_KEY_FILE
The instructions at
https://cloud.google.com/datastore/docs/getstarted/start_python/ lead
one to create a JSON keyfile, but the code didn't actually accept that.
This is a quick hack that just looks at the filename to decide whether
it is a raw key (PKCS#12?) or a JSON-wrapped one, and extract the key
from the JSON file if necessary.