feat: add optional end-to-end checksum support for http 1.1 requests (#261)
Update RemoteRpc to optionally support end-to-end checksums. If enabled,
when sending a request to datastore a checksum will be calculated and 
included as a header of the request. When receiving a response from
datastore and the checksum header is present, a checksum of the payload
will be computed and validated to match the values specified in the
header. If validation of the response fails, an IOException will be thrown
to signal possible payload corruption.

By default, checksum validation is not enabled. To enable checksum validation
the environment variable `GOOGLE_CLOUD_DATASTORE_HTTP_ENABLE_E2E_CHECKSUM`
should be set to `true`.